requirements.txt
# EquiShard (InvestSaaS) Dependencies
# Core Framework
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# FastAPI Query Layer
fastapi>=0.109,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic>=2.5,<3.0
pydantic-settings>=2.1,<3.0
python-multipart>=0.0.6
PyJWT>=2.8,<3.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Production Server
gunicorn>=21.0,<22.0

# Utilities
python-dotenv>=1.0,<2.0
